Create a 'read-only' group for the tests
Create a 'read-only' group for the unit test that can be run publicly
with a public API key.
Also not using JSON output for the tests anymore as this was making it
to odifficult to switch between testing in QA and PROD. Now unit tests
check for field presence and some known values.