Improved convertBool() function
The function convertBool() now takes 2 params, an array of arguments and
the key (the boolean entry) to convert. This needs to be done for cases
where the key is not defined. Looks like some version of PHP let you
pass an 'undefined' as a parameter to a function but some (older)
version will throw an error. So now passing array and key as 2 separate
parameters to the function.